Add api for exclusion list
